vehicle/requirements.txt
alembic==1.15.2
blinker==1.9.0
cffi==1.17.1
click==8.2.0
cryptography==45.0.1
Flask==3.1.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pycparser==2.22
PyMySQL==1.1.0
pytest==8.3.5
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
